vietfuture/backend-vietfuture/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
apscheduler==3.10.4
email-validator==2.1.1
python-dotenv==1.0.1
httpx==0.28.1
google-genai==1.11.0
groq>=0.12.0
requests==2.32.3
aiofiles==24.1.0
qrcode==7.4.2
Pillow>=10.0.0
pywebpush>=1.14.0
